+++ src/Form/Type/InvoiceCalculatorType.php
@@ -9,7 +9,7 @@
 
 namespace App\Form\Type;
 
-use App\Invoice\ServiceInvoice;
+use App\Invoice\InvoiceService;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\OptionsResolver\OptionsResolver;
@@ -19,7 +19,7 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
  */
 final class InvoiceCalculatorType extends AbstractType
 {
-    public function __construct(private ServiceInvoice $service)
+    public function __construct(private InvoiceService $service)
     {
     }
